+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
  - Re-export windows::runtime::Result
+ - Add support for scenarios
 
 ## 0.5.0
  - Updated dependencies